fix: change Task and TaskList position fields from int to float (#10)
The Favro API returns float values for task positions to support
fractional positioning (e.g., inserting between positions 1 and 2
yields 1.5). This was causing Pydantic validation errors.

Fixes #8

+++ src/favro_mcp/api/models.py
@@ -169,7 +169,7 @@ class TaskList(BaseModel):
     organization_id: str = Field(alias="organizationId")
     card_common_id: str = Field(alias="cardCommonId")
     name: str
-    position: int
+    position: float
 
 
 class Task(BaseModel):
@@ -181,4 +181,4 @@ class Task(BaseModel):
     card_common_id: str = Field(alias="cardCommonId")
     name: str
     completed: bool
-    position: int
+    position: float
